Slot Machine Pet Project
Testing myself with a challenge! The initial commit for this project contains the files needed to run the Console App. However this is a functional "game" of chance, the symbols displayed are hard-coded as Emojies. For a better experience playing around with it I would recommend using Windows Terminal, and a Console Prompt profile where the font is set to "Segoe UI Emoji" or other Emoji supporting fonts.